requirements.txt
streamlit==1.32.2
ultralytics==8.1.34
opencv-python-headless==4.9.0.80
numpy<2
pillow
